(refactor) topbar logo
