+++ package.json
@@ -38,12 +38,7 @@
     "eslintConfig": {
         "extends": "react-app"
     },
-    "husky": {
-        "hooks": {
-            "pre-commit": "npm run lint",
-            "pre-push": "npm run lint && npm run test-once"
-        }
-    },
+
     "browserslist": [
         ">0.2%",
         "not dead",
